Reimplements platforms, and the Chiseling (#1163)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Ports:
- Partial https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2166
- Ports Ratwood inhand sprites for bundles

Changes:
- Adds a chisel, which lets you convert stone into blocks, similar to
planks
- New masonry recipes that require blocks
- Floors can no longer be built over openspaces. Instead, gaps and water
turfs can be crossed using "platforms". Wood uses plank, stone uses
blocks, and twig requires a rope.
- Constructed floors can be broken with cmode on. Watch your step if you
miss your swing
- Artificers have lumberjacking instead of carpentry
- Stone lore no longer adds additional damage to the rock when used as a
weapon. Throwing damage is still buffed.
- Adds mason pilgrim
- Bug fixes

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Mason maxxing. Death to skybridgers.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->