Sprite changes to summonable's drops (#1281)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
This changes some sprites in natural.dm, particularly those dropped by
summons,
see details for sprite comparison.
<details> 
sprites before:

![Screenshot 2025-04-01
021530](https://github.com/user-attachments/assets/e4cc3b56-dcc8-48ce-a713-d88a2b3a1c0c)

sprites after:

![Screenshot 2025-04-01
021514](https://github.com/user-attachments/assets/f78c48c2-5d5a-423f-b92f-3df0c62d5805)

</details>
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->