Racelock Adjustments: The Sequel (#1165)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

The systemic racism update!

Reworks/organizes many of the race locks present in the current game.
Some only had a few changes, some had a lot.
- Organized defines for most circumstances.
- `RACES_PLAYER_ALL` - Previously `ALL_PLAYER_RACES_BY_NAME`, works more
in line with the others.
- `RACES_PLAYER_NONDISCRIMINATED` - Humen, Half-Elf, Elf, Dwarf,
Aasimar. Used for guards, nobles, priest and templars.
- ``RACES_PLAYER_NONHERETICAL`` - Humen, Rakshari, Half-Elf, Elf,
Aasimar. Used for lesser religious roles such as acolyte, churchling,
gravedigger, and cleric.
- `RACES_PLAYER_NONEXOTIC`- All races but Kobold and Rakshari. Used for
dungeoneer, physicker, feldsher, apothecary, matron, etc.
- `RACES_PLAYER_GRENZ` - Dwarf, Humen, Aasimar. Used for all Grenz
roles.
- `RACES_PLAYER_ELF` - Elf, Half-Elf. Currently only used for Black oak.
Will see use in Rosewood update.
- `RACES_PLAYER_ELF_D` - Elf, Half-Elf, Dark Elf. Thought this would be
used more but it wasn't, only on Sentinel. Might see use in the future?
- `RACES_PLAYER_PATREON` - Kobold. I just defined it so it was there,
dunno it'll see use in the future.

There are too many minute changes for me to express here. Here are the
highlights.

- **Royal Family**
- Added dwarves, per Ook's request. Vanderlin is a melting pot for
races, and so this was deemed acceptable for it.
 
- **Nobility**
  - Nobles have had their lesser races removed.
 
- **Pilgrim/Towner**
- General inconsistencies between towner roles and their pilgrim
variant. Changed where deemed necessary.

- **Church**
- Removed Dark Elves / Tieflings / Horcs from acolytes/gravetender, as
these are heretical races.
  - Removed elves from inquisitor, as they are Grenz.

- **Rakshari**
  - Rakshari got the short end when their roles were added.
- Given adventurer roles of warrior, ranger, mage/sorceress, Qatil, and
Treasure Hunter
  - Access to all Zybantine roles.

- **Kobolds** 
  - Removed feldsher, squire, wandering physicker.
  - Added most pilgrim roles and mercenary underdweller.

- **Migrant Waves**
  - Some confusing things in here.
  - Jester troupe unlocked to all races
- Prisoner wave unlocked to all races. Guards races held to guard
standard.
  - Deprived unlocked to all races
  - Heartfelt can be most races now
- Removed Tief, Drow and Horc from Zybantine Emir/Amirah. Furusiyya open
to elves, dwarves, and Rakshari. Slave guard opened to Rakshari. Qatil
opened to Rakshari. Emir/Amirah can now be children (Emir/Amirah is a
catch-all for prince/princess).


## Why It's Good For The Game
I feel pretty justified in these changes. There are some minor ones I
glossed over, mostly minor horc stuff. Feedback is appreciated.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>