(FIX) Fixes Templar Knight Voiceline, re-implements Knight voicelines to Knight. (#1352)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Yeah turns out there was a fucking typo preventing Templars from using
the Knight voicelines. It was humen. It's meant to be human. God fucking
damnit.

Also re-implements the Royal Guards having the knight voicepack. I don't
know why this was removed. Practically only migrant waves use it.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->